@@ -19,19 +19,20 @@
   },
   "homepage": "https://github.com/mercadolibre/fury-core-actions#readme",
   "dependencies": {
-    "@actions/core": "^1.10.0",
-    "@actions/github": "^2.1.1",
+    "@actions/core": "^1.10.1",
+    "@actions/github": "^5.1.1",
     "@types/lodash": "^4.14.74",
-    "dayjs": "^1.8.26",
     "semver": "^7.5.4"
   },
   "devDependencies": {
     "@babel/types": "^7.14.0",
-    "@zeit/ncc": "^0.20.5",
+    "@types/node": "^20.6.0",
+    "@types/semver": "^7.5.1",
+    "@vercel/ncc": "^0.38.0",
     "eslint": "^6.3.0",
     "husky": "^4.2.5",
     "jest": "^26.0.0",
-    "typescript": "^3.4.2"
+    "typescript": "^5.2.2"
   },
   "husky": {
     "hooks": {
